+++ input/parameters/lithium-ion/electrolytes/lipf6_Marquis2019/electrolyte_conductivity_Capiglia1999.py
@@ -1,4 +1,4 @@
-import autograd.numpy as np
+from pybamm import exp
 
 
 def electrolyte_conductivity_Capiglia1999(c_e, T, T_inf, E_k_e, R_g):
@@ -38,6 +38,6 @@ def electrolyte_conductivity_Capiglia1999(c_e, T, T_inf, E_k_e, R_g):
         + 0.1554 * (c_e / 1000) ** 3
     )
 
-    arrhenius = np.exp(E_k_e / R_g * (1 / T_inf - 1 / T))
+    arrhenius = exp(E_k_e / R_g * (1 / T_inf - 1 / T))
 
     return sigma_e * arrhenius