+++ pybamm/parameters/geometric_parameters.py
@@ -36,14 +36,12 @@ a_n_dim = pybamm.Parameter("Negative electrode surface area density [m-1]")
 a_p_dim = pybamm.Parameter("Positive electrode surface area density [m-1]")
 R_n = pybamm.Parameter("Negative particle radius [m]")
 R_p = pybamm.Parameter("Positive particle radius [m]")
-b_n = pybamm.Parameter("Negative electrode Bruggeman coefficient")
-b_s = pybamm.Parameter("Separator Bruggeman coefficient")
-b_p = pybamm.Parameter("Positive electrode Bruggeman coefficient")
-b = pybamm.Concatenation(
-    pybamm.FullBroadcast(b_n, ["negative electrode"], "current collector"),
-    pybamm.FullBroadcast(b_s, ["separator"], "current collector"),
-    pybamm.FullBroadcast(b_p, ["positive electrode"], "current collector"),
-)
+b_e_n = pybamm.Parameter("Negative electrode Bruggeman coefficient (electrolyte)")
+b_e_s = pybamm.Parameter("Separator Bruggeman coefficient (electrolyte)")
+b_e_p = pybamm.Parameter("Positive electrode Bruggeman coefficient (electrolyte)")
+b_s_n = pybamm.Parameter("Negative electrode Bruggeman coefficient (electrode)")
+b_s_s = pybamm.Parameter("Separator Bruggeman coefficient (electrode)")
+b_s_p = pybamm.Parameter("Positive electrode Bruggeman coefficient (electrode)")
 
 # --------------------------------------------------------------------------------------
 "Dimensionless Parameters"