+++ pybamm/solvers/idaklu_solver.py
@@ -48,7 +48,7 @@ class IDAKLUSolver(pybamm.DaeSolver):
     def set_atol_by_variable(self, variables_with_tols, model):
         """
         A method to set the absolute tolerances in the solver by state variable.
-        This method modifies self._atol.
+        This method attaches a vector of tolerance to the model. (i.e. model.atol)
 
         Parameters
         ----------
@@ -61,15 +61,15 @@ class IDAKLUSolver(pybamm.DaeSolver):
         """
 
         size = model.concatenated_initial_conditions.size
-        self._check_atol_type(size)
+        atol = self._check_atol_type(self._atol, size)
         for var, tol in variables_with_tols.items():
             variable = model.variables[var]
             if isinstance(variable, pybamm.StateVector):
-                self.set_state_vec_tol(variable, tol)
+                atol = self.set_state_vec_tol(atol, variable, tol)
             elif isinstance(variable, pybamm.Concatenation):
                 for child in variable.children:
                     if isinstance(child, pybamm.StateVector):
-                        self.set_state_vec_tol(child, tol)
+                        atol = self.set_state_vec_tol(atol, child, tol)
                     else:
                         raise pybamm.SolverError(
                             """Can only set tolerances for state variables
@@ -81,7 +81,9 @@ class IDAKLUSolver(pybamm.DaeSolver):
                     concatenations of state variables"""
                 )
 
-    def set_state_vec_tol(self, state_vec, tol):
+        model.atol = atol
+
+    def set_state_vec_tol(self, atol, state_vec, tol):
         """
         A method to set the tolerances in the atol vector of a specific
         state variable. This method modifies self._atol
@@ -94,37 +96,43 @@ class IDAKLUSolver(pybamm.DaeSolver):
             The tolerance value
         """
         slices = state_vec.y_slices[0]
-        self._atol[slices] = tol
+        atol[slices] = tol
+        return atol
 
-    def _check_atol_type(self, size):
+    def _check_atol_type(self, atol, size):
         """
         This method checks that the atol vector is of the right shape and
         type.
 
         Parameters
         ----------
+        atol: double or np.array or list
+            Absolute tolerances. If this is a vector then each entry corresponds to
+            the absolute tolerance of one entry in the state vector.
         size: int
             The length of the atol vector
         """
 
-        if isinstance(self._atol, float):
-            self._atol = self._atol * np.ones(size)
-        elif isinstance(self._atol, list):
-            self._atol = np.array(self._atol)
-        elif isinstance(self._atol, np.ndarray):
+        if isinstance(atol, float):
+            atol = atol * np.ones(size)
+        elif isinstance(atol, list):
+            atol = np.array(atol)
+        elif isinstance(atol, np.ndarray):
             pass
         else:
             raise pybamm.SolverError(
                 "Absolute tolerances must be a numpy array, float, or list"
             )
 
-        if self._atol.size != size:
+        if atol.size != size:
             raise pybamm.SolverError(
                 """Absolute tolerances must be either a scalar or a numpy arrray
                 of the same shape at y0"""
             )
 
-    def integrate(self, residuals, y0, t_eval, events, mass_matrix, jacobian):
+        return atol
+
+    def integrate(self, residuals, y0, t_eval, events, mass_matrix, jacobian, model):
         """
         Solve a DAE model defined by residuals with initial conditions y0.
 
@@ -146,6 +154,8 @@ class IDAKLUSolver(pybamm.DaeSolver):
             A function that takes in t and y and returns the Jacobian. If
             None, the solver will approximate the Jacobian.
             (see `SUNDIALS docs. <https://computation.llnl.gov/projects/sundials>`).
+        model : :class:`pybamm.BaseModel`
+            The model whose solution to calculate.
         """
 
         if jacobian is None:
@@ -154,9 +164,13 @@ class IDAKLUSolver(pybamm.DaeSolver):
         if events is None:
             pybamm.SolverError("KLU requires events to be provided")
 
+        try:
+            atol = model.atol
+        except AttributeError:
+            atol = self._atol
+
         rtol = self._rtol
-        self._check_atol_type(y0.size)
-        atol = self._atol
+        atol = self._check_atol_type(atol, y0.size)
 
         if jacobian:
             jac_y0_t0 = jacobian(t_eval[0], y0)